\name{.}
\alias{.}
\title{Apply Function Call to Zelig Objects Result Slot...}
\description{Apply Function Call to Zelig Objects Result Slot}
\value{the result of applying the expression to the 
`result' slot of the first parameter}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{expr}{an expression to lazily-evaluate}
\item{envir}{ignored until later versions}}
