\name{.copy.templates}
\alias{.copy.templates}
\title{creates _necessary_ files for a model...}
\usage{.copy.templates(model, pkg, path)}
\description{creates _necessary_ files for a model
param model a character-string specifying the name of the model
param pkg a character-string specifying the name of the package
param path a character-string specifying the base path to the package
return nothing}
