\name{.generate.functions}
\alias{.generate.functions}
\title{generate zelig2, qi.}
\usage{.generate.functions(model, env)}
\description{generate zelig2, qi., and param. functions for model
in a particular environment
param model a character-string specifying the name of a non-existent Zelig model
param env an environment 
return invisible environment variable (with stored functions)}
