\name{.get.list.as.text}
\alias{.get.list.as.text}
\title{Convert Character-Strings into Human-Readable Lists...}
\usage{.get.list.as.text(..., final.comma=FALSE)}
\description{Convert Character-Strings into Human-Readable Lists
This functions converts its parameters into a human-readable and
grammatically correct series.}
\value{a comma delineated string}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{...}{character-vectors and list of characters}
\item{final.comma}{whether to add the final comma to the series. Grammatical
correctness is debateable}}
