\name{.GetModelCitationTex}
\alias{.GetModelCitationTex}
\title{Get a Text-Block of Citation Information about a Zelig Model...}
\usage{.GetModelCitationTex(model.name)}
\description{Get a Text-Block of Citation Information about a Zelig Model}
\note{This function is strictly used internally by Zelig}
\value{a block of text giving a human readable (and APA compliant) block
citation text}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{model.name}{the name of a Zelig model}}
