\name{.intersection}
\alias{.intersection}
\title{Compute the Intersection of Two Sets...}
\usage{.intersection(a, b, unique=TRUE)}
\description{Compute the Intersection of Two Sets}
\note{This function is used internally by Zelig}
\value{the intersection of a and b}
\arguments{\item{a}{a vector}
\item{b}{a vector}
\item{unique}{a boolean determining whether a intersect b will contain only
unique elements}}
