\name{.make.description}
\alias{.make.description}
\title{make a description file for a specific package...}
\usage{.make.description(pkg, author, email, depends, url, path=".")}
\description{make a description file for a specific package
param pkg a character-string specifying the name of the package
param author a vector of strings specifying the names of contributors
param email a character-string specifying the email of the maintainer
param depends a vector of strings specifying package dependencies
param url - ignored -
param path a character-string specifying the location of the package
return nothing}
