\name{.make.package.R}
\alias{.make.package.R}
\title{Make pkg-package...}
\usage{.make.package.R(pkg, author, email, depends, url, path=".")}
\description{Make \code{pkg}-package.R File for Roxygen Compliancy}
\value{NULL}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{pkg}{the package name}
\item{author}{a vector of characters specifying the authors of the Zelig
models}
\item{email}{the email address of the package's maintainer}
\item{depends}{a vector specifying package dependencies}
\item{URL}{specifying the package's website}
\item{path}{location of the package}}
