\name{.prefix}
\alias{.prefix}
\title{Append a Prefix to a Character String...}
\usage{.prefix(name, envir, prefix="zelig", sep=".")}
\description{Append a Prefix to a Character String}
\note{This function is exclusively used internally by Zelig}
\arguments{\item{name}{a character-string specifying the name of a variable}
\item{envir}{an environment variable to search}
\item{prefix}{a character-string to prefix the string with
this is applied until the name is unique}
\item{sep}{a character-string that separates prefix and name}}
