\name{.reduceFurther}
\alias{.reduceFurther}
\title{Further Reduce Formulas in Reduced Form...}
\usage{.reduceFurther(f)}
\description{Further Reduce Formulas in Reduced Form
Given a formula in a reduced form, output the most reduced one.}
\value{an even-more reduced formula}
\author{Ferdinand Alimadhi, Kosuke Imai, and Olivia Lau}
\arguments{\item{f}{a formula in reduced form}}
