\name{.run}
\alias{.run}
\title{Run Model Fitting Function in Encapsulated Environment...}
\usage{.run(...)}
\description{Run Model Fitting Function in Encapsulated Environment}
\note{This function is exclusively used internally Zelig}
\value{the object ran in an insulated environment}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{..1}{a zelig.call object}
\item{..2}{an environment object}}
