\name{.store.values}
\alias{.store.values}
\title{Store Named Values in an Newly Created Environment...}
\usage{.store.values(params)}
\description{Store Named Values in an Newly Created Environment}
\note{This function is exclusively used internally by Zelig}
\value{a list containing:}
\arguments{\item{params}{a list of parameters to create a Zelig Call}}
