\name{.substitute.expressions}
\alias{.substitute.expressions}
\title{.substitute.expressions}
\usage{.substitute.expressions(.file, ...)}
\note{This function fails if passed non-alphanumeric variable names. In
particular, the parameters cannot contain periods, etc.}
\value{a character-string}
\arguments{\item{.file}{the name of the file to replace}
\item{...}{}}
