\name{.ZeligDescribeModel}
\alias{.ZeligDescribeModel}
\title{Describe a Zelig Model...}
\usage{.ZeligDescribeModel(model.name, ...)}
\description{Describe a Zelig Model}
\value{a 'description' object containing citation information}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{model.name}{}
\item{...}{ignored parameters}}
