\name{.ZeligModelCategories}
\alias{.ZeligModelCategories}
\title{Get a List of Categories for Describing Zelig Models...}
\description{Get a List of Categories for Describing Zelig Models}
\note{This feature is being deprecated, as original functionality with the
Dataverse Project \url{thedata.org} is being reevaluated.}
\value{a list of character-string specifying legal category types (as the
keys of the list) and their human-counterparts (as the values)}
