\name{[.zframe}
\alias{[.zframe}
\title{Extract Subsetted Data-Frames from zframe Objects...}
\usage{[.zframe(z, ...)}
\description{Extract Subsetted Data-Frames from \code{zframe} Objects
\code{zframe} objects are exclusively used internally within Zelig; they 
facilitate the process of subsetting data-frames for the \code{mi} function.}
\note{The replacement operator is undefined for \code{zframe} objects.}
\value{a subsetted data-frame}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{z}{a 'zframe' object used to describe multiple subsetted data-sets}
\item{...}{parameters to be passed to the default extraction operator}}
