\name{[[.zelig}
\alias{[[.zelig}
\title{Extract a Value from a Fitted Model Object (Wrapped by Zelig)...}
\description{Extract a Value from a Fitted Model Object (Wrapped by Zelig)}
\value{contents of the specified slots}
\author{Matt Owen \emph{mowen@iq.harvard.edu}}
\arguments{\item{z}{an object of type 'zelig'}
\item{slot}{a character-string specifying the slot to extract from the fitted
model object}
\item{...}{subsequent slots to extract from the fitted model object}}
