\name{alpha}
\alias{alpha}
\title{Extract ancillary parameters from...}
\usage{alpha(param)}
\description{Extract ancillary parameters from
`parameters' objects}
\value{the ancillary parameters \emph{specified} for
the statistical model}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{param}{a `parameters' object}}
