\name{as.bootlist}
\alias{as.bootlist}
\title{Convert of Vector of Bootstrapped Parameters to a List-style Boot Object}
\usage{
  as.bootlist(bootstraps, lengths, names)
}
\arguments{
  \item{bootstraps}{...}

  \item{lengths}{...}
}
\value{
  ...
}
\description{
  This inverts the ``as.bootvector'' function, and returns
  a list containing the slots ``alpha'' and ``beta''.
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

