\name{as.data.frame.zframe}
\alias{as.data.frame.zframe}
\title{Extract the Data-Frame...}
\usage{as.data.frame.zframe(x, row.names, optional=FALSE, ...)}
\description{Extract the Data-Frame}
\note{This method is only intended for internal use by Zelig; its
functionality lacks enough polish for interactive use by users or use by
other software packages (without reading this warning).}
\value{the corresponding subsetted data-frame}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{x}{a 'zframe' object}
\item{row.names}{ignored}
\item{optional}{ignored}
\item{...}{ignored}}
