\name{as.description.description}
\alias{as.description.description}
\title{description -> description}
\usage{
  as.description.description(descr, ...)
}
\arguments{
  \item{descr}{an object of type 'description'}

  \item{...}{ignored}
}
\value{
  the same object
}
\description{
  description -> description
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

