\name{as.description.list}
\alias{as.description.list}
\title{list -> description}
\usage{
  \method{as.description}{list}(descr, ...)
}
\arguments{
  \item{descr}{a list}

  \item{...}{ignored}
}
\value{
  an object of type 'description'
}
\description{
  Convert list into a description object.
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

