\name{as.parameters.default}
\alias{as.parameters.default}
\title{??? -> parameters}
\usage{
  as.parameters.default(params, num = NULL, ...)
}
\arguments{
  \item{params}{any non-supported data-type}

  \item{num}{an integer specifying the number of
  simulations to compute}

  \item{...}{ignored}
}
\value{
  the object passed in
}
\description{
  ??? -> parameters
}
\note{
  This function should be deprecated.
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

