\name{as.parameters.list}
\alias{as.parameters.list}
\title{list -> parameters...}
\usage{as.parameters.list(params, num, ...)}
\description{list -> parameters}
\details{The list may contain: 'link', 'linkinv', 'family', 'alpha', and
'simulations' keys.}
\value{an object of type `parameters'}
\seealso{as.parameters}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{params}{a list object}
\item{num}{an integer specifying the number of simulations
to be taken}
\item{...}{ignored parameters}}
