\name{as.parameters.parameters}
\alias{as.parameters.parameters}
\title{parameters -> parameters
This is merely an identity function when casting 'parameters' objects into
'parameters'.}
\usage{
  as.parameters.parameters(params, ...)
}
\arguments{
  \item{params}{a parameters object}

  \item{...}{ignored parameters}
}
\value{
  the same parameter object
}
\description{
  parameters -> parameters This is merely an identity
  function when casting 'parameters' objects into
  'parameters'.
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

