\name{as.summarized.list}
\alias{as.summarized.list}
\title{list -> summarized.qi}
\usage{
  as.summarized.list(x, ...)
}
\arguments{
  \item{x}{a list}

  \item{...}{ignored parameters}
}
\value{
  a 'summarized.qi' object
}
\description{
  list -> summarized.qi
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

