\name{as.summarized.summarized.qi}
\alias{as.summarized.summarized.qi}
\title{summarized...}
\usage{as.summarized.summarized.qi(x, ...)}
\description{summarized.qi -> summarized.qi
This is an identity operation}
\value{the same 'summarized.qi' object}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{x}{an object of type 'summarized.qi'}
\item{...}{ignored parameters}}
