\name{bootstrap.default}
\alias{bootstrap.default}
\title{Produce Boot-strapped Parameters for a Statistical Model}
\usage{
  bootstrap.default(obj, num, ...)
}
\arguments{
  \item{obj}{1}

  \item{num}{an integer specifying the number of
  simulations to produce}

  \item{...}{}
}
\value{
  a list with the ``link'', ``linkinv'' and ``family''
  slots set
}
\description{
  ...
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

