\name{callsystemfit}
\alias{callsystemfit}
\title{Fit Multi-stage Linear Models}
\usage{
  callsystemfit(formula, data, method, inst = NULL, ...)
}
\arguments{
  \item{formula}{a ``formula'' object}

  \item{data}{a ``data.frame'' object}

  \item{method}{a character-string}

  \item{inst}{instrument parameters}

  \item{...}{other parameters passed to systemfit}
}
\description{
  Fit Multi-stage Linear Models
}

