\name{cite}
\alias{cite}
\title{Citation information for a 'description' object}
\usage{
  cite(descr)
}
\arguments{
  \item{descr}{an object of type 'description'}
}
\value{
  a character-string giving citation info
}
\description{
  Citation information for a 'description' object
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

