\name{cmvglm}
\alias{cmvglm}
\title{cmvglm}
\usage{
  cmvglm(formula, model, ndim, data = NULL, fact = NULL)
}
\arguments{
  \item{formula}{a formula}

  \item{model}{the names of the Zelig model}

  \item{ndim}{the number of dimensions in the statistical
  model}

  \item{data}{a data-frame}

  \item{fact}{???}
}
\description{
  cmvglm
}
\author{
  Kosuke Imai and Olivia Lau
}

