\name{combine}
\alias{combine}
\title{Produce All Combinations of a Set of Lists...}
\usage{combine(...)}
\description{Produce All Combinations of a Set of Lists}
\note{This function is used internall by the 'mi' constructors in order to
produce the complete set of combinations of data-frames and factors by
to subset the data-frames.}
\value{all the combinations of the lists with repetition}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{...}{a set of lists to mix together}}
