\name{constructDataFrame}
\alias{constructDataFrame}
\title{Construct Data Frame
Construct and return a tiny (single-row) data-frame from a larger data-frame,
a list of specified values, and a formula}
\usage{
  constructDataFrame(data, specified)
}
\arguments{
  \item{data}{a ``data.frame'' that will be used to create
  a small design matrix}

  \item{specified}{a list with key-value pairs that will be
  used to explicitly set several values}
}
\value{
  a ``data.frame'' containing a single row
}
\description{
  Construct Data Frame Construct and return a tiny
  (single-row) data-frame from a larger data-frame, a list
  of specified values, and a formula
}

