\name{constructDesignMatrix}
\alias{constructDesignMatrix}
\title{Construct Design Matrix from
Construct and return a design matrix based on a tiny data-frame (single-row).}
\usage{
  constructDesignMatrix(data, formula)
}
\arguments{
  \item{data}{a ``data.frame'' (preferably single-rowed)
  that will be used to create a small design matrix}

  \item{formula}{a formula, whose predictor variables will
  be used to create a design matrix}
}
\value{
  a design (model) matrix
}
\description{
  Construct Design Matrix from Construct and return a
  design matrix based on a tiny data-frame (single-row).
}

