\name{constuctDataFrame}
\alias{constuctDataFrame}
\title{Construct Data Frame
Construct and return a tiny (single-row) data-frame from a larger data-frame,
a list of specified values, and a formula}
\usage{
  constuctDataFrame(data, specified, formula)
}
\arguments{
  \item{data}{...}

  \item{specified}{...}

  \item{formula}{...}
}
\value{
  ...
}
\description{
  Construct Data Frame Construct and return a tiny
  (single-row) data-frame from a larger data-frame, a list
  of specified values, and a formula
}

