\name{describe.default}
\alias{describe.default}
\title{Default describe function for an arbitrary model
This method exists solely as a backup when an author does not contribute a
'describe' function for their model}
\usage{
  describe.default(...)
}
\arguments{
  \item{...}{dummy parameters purely to cast the correct
  object. That is, the parameters of the function should
  not BE referenced specifically}
}
\value{
  a list to be processed by \code{as.description}
}
\description{
  Default describe function for an arbitrary model This
  method exists solely as a backup when an author does not
  contribute a 'describe' function for their model
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

