\name{describe.factor.bayes}
\alias{describe.factor.bayes}
\title{Describe the factor.bayes Zelig Model}
\usage{
  describe.factor.bayes(...)
}
\arguments{
  \item{...}{ignored parameters}
}
\value{
  a list specifying author, title, etc. information
}
\description{
  Describe the factor.bayes Zelig Model
}

