\name{describe.gamma}
\alias{describe.gamma}
\title{Describe the \code{gamma} model to Zelig}
\usage{
  describe.gamma(...)
}
\arguments{
  \item{...}{ignored parameters}
}
\value{
  a list of important information
}
\description{
  Describe the \code{gamma} model to Zelig
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

