\name{describe.logit.bayes}
\alias{describe.logit.bayes}
\title{Describe the logit.bayes Zelig Model}
\usage{
  describe.logit.bayes(...)
}
\arguments{
  \item{...}{ignored parameters}
}
\value{
  a list specifying author, title, etc. information
}
\description{
  Describe the logit.bayes Zelig Model
}

