\name{describe.logit}
\alias{describe.logit}
\title{Describe a `logit' model to Zelig...}
\usage{describe.logit(...)}
\description{Describe a `logit' model to Zelig}
\value{a list to be processed by `as.description'}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{...}{ignored parameters}}
