\name{describe.mlogit.bayes}
\alias{describe.mlogit.bayes}
\title{Describe the mlogit.bayes Zelig Model}
\usage{
  describe.mlogit.bayes(...)
}
\arguments{
  \item{...}{ignored parameters}
}
\value{
  a list specifying author, title, etc. information
}
\description{
  Describe the mlogit.bayes Zelig Model
}

