\name{describe.normal.bayes}
\alias{describe.normal.bayes}
\title{Describe the normal.bayes Zelig Model}
\usage{
  describe.normal.bayes(...)
}
\arguments{
  \item{...}{ignored parameters}
}
\value{
  a list specifying author, title, etc. information
}
\description{
  Describe the normal.bayes Zelig Model
}

