\name{describe.normal}
\alias{describe.normal}
\title{Describe the normal model to Zelig...}
\usage{describe.normal(...)}
\description{Describe the \code{normal} model to Zelig}
\value{a list to be processed by `as.description'}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{...}{ignored parameters}}
