\name{describe.normal.survey}
\alias{describe.normal.survey}
\title{Describe a \code{normal.survey} Citation to Zelig}
\usage{
  describe.normal.survey(...)
}
\arguments{
  \item{...}{ignored parameters}
}
\value{
  a list to be processed by \code{as.description}
}
\description{
  Describe a \code{normal.survey} Citation to Zelig
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

