\name{describe.oprobit.bayes}
\alias{describe.oprobit.bayes}
\title{Describe the oprobit.bayes Zelig Model}
\usage{
  describe.oprobit.bayes(...)
}
\arguments{
  \item{...}{ignored parameters}
}
\value{
  a list specifying author, title, etc. information
}
\description{
  Describe the oprobit.bayes Zelig Model
}

