\name{describe.poisson.bayes}
\alias{describe.poisson.bayes}
\title{Describe the poisson.bayes Zelig Model}
\usage{
  describe.poisson.bayes(...)
}
\arguments{
  \item{...}{ignored parameters}
}
\value{
  a list specifying author, title, etc. information
}
\description{
  Describe the poisson.bayes Zelig Model
}

