\name{describe.poisson}
\alias{describe.poisson}
\title{Describe the `poisson' model to Zelig...}
\usage{describe.poisson(...)}
\description{Describe the `poisson' model to Zelig}
\value{a list to be processed by `as.description'}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{...}{ignored parameters}}
