\name{describe.probit.bayes}
\alias{describe.probit.bayes}
\title{Describe the probit.bayes Zelig Model}
\usage{
  describe.probit.bayes(...)
}
\arguments{
  \item{...}{ignored parameters}
}
\value{
  a list specifying author, title, etc. information
}
\description{
  Describe the probit.bayes Zelig Model
}

