\name{describe.probit}
\alias{describe.probit}
\title{Describe the `probit' model to Zelig...}
\usage{describe.probit(...)}
\description{Describe the `probit' model to Zelig}
\value{a list to be processed by `as.description'}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{...}{ignored parameters}}
