\name{describe}
\alias{describe}
\title{Method to describe a model to Zelig...}
\usage{describe(...)}
\description{Method to describe a model to Zelig}
\value{a list to be processed by `as.description'}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{...}{parameters which are typically ignored}}
