\name{describe.threesls}
\alias{describe.threesls}
\title{Describe a `threesls' model to Zelig}
\usage{
  \method{describe}{threesls}(...)
}
\arguments{
  \item{...}{ignored parameters}
}
\value{
  a list to be processed by `as.description'
}
\description{
  Describe a `threesls' model to Zelig
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

