\name{describe.tobit}
\alias{describe.tobit}
\title{Describe a ``tobit'' model to Zelig}
\usage{
  \method{describe}{tobit}(...)
}
\arguments{
  \item{...}{ignored parameters}
}
\value{
  a list to be processed by `as.description'
}
\description{
  Describe a ``tobit'' model to Zelig
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

