\name{describe.twosls}
\alias{describe.twosls}
\title{Describe a `twosls' model to Zelig}
\usage{
  \method{describe}{twosls}(...)
}
\arguments{
  \item{...}{ignored parameters}
}
\value{
  a list to be processed by `as.description'
}
\description{
  Describe a `twosls' model to Zelig
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

