\name{describe.zelig}
\alias{describe.zelig}
\title{Get Description Object Used to Cite this Zelig Model...}
\usage{describe.zelig(object, ...)}
\description{Get Description Object Used to Cite this Zelig Model}
\note{This function should be reevaluated in design, since 'description'
objects are exclusively used internally. In particular, this method would
be more useful to users as a 'cite' method.}
\value{a 'description' object used internally to produce citation text}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{object}{a 'zelig' object}
\item{...}{ignored parameters}}
