\name{eval.in}
\alias{eval.in}
\title{Attach Environment and Evaluate Expression...}
\description{Attach Environment and Evaluate Expression}
\value{evaluation of ..1}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{..1}{an expression to evaluate}
\item{..2}{the environment to evaluate the expression in}}
