\name{get.mi.labels}
\alias{get.mi.labels}
\title{Get List of Labels from MI object...}
\usage{get.mi.labels(obj, ...)}
\description{Get List of Labels from MI object}
\value{a vector of character-strings detailing - in human readable format -
the titles of each}
\arguments{\item{obj}{an mi object}
\item{...}{ignored (for now) parameters}}
