\name{GetObject}
\alias{GetObject}
\title{Extract the fitted model object from the Zelig object...}
\usage{GetObject(obj)}
\description{Extract the fitted model object from the Zelig object}
\value{the fitted model object}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{obj}{an object of type `zelig'}}
