\name{getPredictorTerms}
\alias{getPredictorTerms}
\title{Get Predictor Terms from Zelig-style Formulae}
\usage{
  getPredictorTerms(x, ...)
}
\arguments{
  \item{x}{a Zelig-style formula ('formula' or 'list')}

  \item{...}{ignored parameters}
}
\value{
  a character-vector or NA
}
\description{
  This function extracts the predictor terms from a
  Zelig-style object.
}
\note{
  This function is used exclusively in the development of
  Zelig-core.
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

