\name{getResponseTerms.list}
\alias{getResponseTerms.list}
\title{Get Response Terms from a List-style Formula...}
\usage{\method{getResponseTerms}{list}(x, ...)}
\description{Get Response Terms from a List-style Formula}
\details{This method gets the response terms from a standard formula}
\value{a character-vector specifying the response terms of the formula}
\author{Matt Owen}
\arguments{\item{x}{a list of formulae}
\item{...}{ignored parameters}}
