\name{GetSlot.zelig}
\alias{GetSlot.zelig}
\title{Extract a slot from a `zelig' object...}
\usage{GetSlot.zelig(z, slot, ...)}
\description{Extract a slot from a `zelig' object}
\value{contents of the specified slots}
\author{Matt Owen \emph{mowen@iq.harvard.edu}}
\arguments{\item{z}{an object of type `zelig'}
\item{slot}{a character-string specifying the slot to 
extract from the fitted model object}
\item{...}{subsequent slots to extract from the fitted
model object}}
