\name{has.zelig2}
\alias{has.zelig2}
\title{Whether an Arbitrary R-package has a Zelig2 Function within Its Namespace...}
\usage{has.zelig2(pkg)}
\description{Whether an Arbitrary R-package has a Zelig2 Function within Its Namespace}
\note{This function is used primarily internally to determine whether a
a package is contributing a function to the Zelig software suite}
\value{whether the package contains any zelig2-functions}
\arguments{\item{pkg}{a character-string representing a package name}}
