\name{help.zelig}
\alias{help.zelig}
\title{Help system for Zelig models}
\usage{
  help.zelig(...)
}
\arguments{
  \item{...}{the help files to look-up}
}
\value{
  results of calling the specific help function
}
\description{
  Help system for Zelig models
}
\author{
  Matt Owen \emph{mowen@iq.harvard.edu}
}

