\name{ignore}
\alias{ignore}
\title{Constructor for the 'ignore' class
This class is included for future use, and is currently
not used in any Zelig model. It is designed for use with
zelig2* functions}
\usage{
  ignore(default = NULL, type = "no pass")
}
\arguments{
  \item{default}{default value}

  \item{type}{ignored parameter}
}
\value{
  an 'ignore' object
}
\description{
  Constructor for the 'ignore' class This class is included
  for future use, and is currently not used in any Zelig
  model. It is designed for use with zelig2* functions
}
\author{
  Matt Owen \emph{mowen@iq.harvard.edu}
}

