\name{is.formula}
\alias{is.formula}
\title{Whether an Object is a Formula...}
\usage{is.formula(x)}
\description{Whether an Object is a Formula}
\details{This is a boolean-check to see whether an object is a formula.}
\note{This will not be shared in the Zelig/ZeligFormulae namespace.}
\value{a logical specifying whether an object is a formula}
\author{Matt Owen}
\arguments{\item{x}{an object}}
