\name{is.qi}
\alias{is.qi}
\title{Test If Value is Interpretable as a QI...}
\usage{is.qi(qi)}
\description{Test If Value is Interpretable as a QI}
\value{a logical specifying whether this value should or should-not
be output}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{qi}{a potential quantity of interest}}
