\name{is.valid.qi.list}
\alias{is.valid.qi.list}
\title{Check If Object Is a List of Valid Quantities of Interest}
\usage{
  is.valid.qi.list(x)
}
\arguments{
  \item{x}{an object to be tested}
}
\value{
  TRUE or FALSE
}
\description{
  Check If Object Is a List of Valid Quantities of Interest
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

