\name{is.zelig.package}
\alias{is.zelig.package}
\title{Wether an Installed R-Pack Depends on Zelig...}
\usage{is.zelig.package(package="")}
\description{Wether an Installed R-Pack Depends on Zelig}
\note{This package was used internally to determine whether an R-package is
Zelig compliant, but is now likely deprecated. This test is useless if not
paired with}
\value{whether this package depends on Zelig}
\arguments{\item{package}{a character-string naming a package}}
