\name{iter.qi}
\alias{iter.qi}
\title{Construct an Iterator from a List of Quantities of Interest...}
\usage{iter.qi(obj, ...)}
\description{Construct an Iterator from a List of Quantities of Interest}
\note{This function is used internally by Zelig}
\value{an iterator}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{obj}{a 'qi' object created by the 'qi' function}
\item{...}{ignored parameters}}
