\name{iter.summarized.qi}
\alias{iter.summarized.qi}
\title{Construct an Iterator from a List of Summarized Quantities of Interest...}
\usage{\method{iter}{summarized.qi}(obj, ...)}
\description{Construct an Iterator from a List of Summarized Quantities of Interest}
\note{This function is used internally by Zelig.}
\value{an iterator}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{obj}{a 'summarized.qi' object}
\item{...}{ignored}}
