\name{length.mi}
\alias{length.mi}
\title{Length Method for 'mi' Objects
Compute the length of the number data-frames to be expected.}
\usage{
  \method{length}{mi}(x)
}
\arguments{
  \item{x}{an object of type 'mi'}
}
\value{
  an integer specifying the number of subsetted
  data-frame's that are contained within within the 'mi'
  object
}
\description{
  Length Method for 'mi' Objects Compute the length of the
  number data-frames to be expected.
}
\note{
  This function is primarily used internally by Zelig.

  This function is exported, since it could be a
  conceivably useful feature for end-users.
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

