\name{link}
\alias{link}
\title{Method for extracting the link function from 'parameters' objects...}
\usage{link(param)}
\description{Method for extracting the link function from 'parameters' objects}
\value{the link function specified by the `param' function for the
given Zelig model}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{param}{a 'parameters' object}}
