\name{linkinv}
\alias{linkinv}
\title{Method for extracting the inverse link function from 'parameters' objects...}
\usage{linkinv(param)}
\description{Method for extracting the inverse link function from 'parameters' objects}
\value{the inverse link function specified by the 'param' function for the
given Zelig model}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{param}{a 'parameters' object}}
