\name{list.zelig.dependent.packages}
\alias{list.zelig.dependent.packages}
\title{Get a List of Packages Installed on the Current Machine that Depend on Zelig}
\usage{
  list.zelig.dependent.packages()
}
\value{
  a character-vector of all zelig-dependent packages on the
  current machine
}
\description{
  Get a List of Packages Installed on the Current Machine
  that Depend on Zelig
}
\note{
  This function is used primarily internally to determine
  whether a a package is contributing a function to the
  Zelig software suite
}

