\name{list.zelig.models}
\alias{list.zelig.models}
\title{List Zelig Models Installed on the Current Machine...}
\usage{list.zelig.models(with.namespace=TRUE)}
\description{List Zelig Models Installed on the Current Machine}
\note{This list is not necessarily complete}
\value{list of all zelig models}
\arguments{\item{with.namespace}{a boolean specifying whether}}
