\name{make.parameters}
\alias{make.parameters}
\title{??? For use with cmvglm}
\usage{
  make.parameters(terms, shape = "vector",
    ancillary = TRUE, eqns = NULL)
}
\arguments{
  \item{terms}{???}

  \item{shape}{???}

  \item{ancillary}{???}

  \item{eqns}{???}
}
\value{
  ???
}
\description{
  ??? For use with cmvglm
}
\author{
  Kosuke Imai and Olivia Lau
}

