\name{Max}
\alias{Max}
\title{Compute the Maximum Value of a Vector}
\usage{
  Max(x, na.rm = NULL)
}
\arguments{
  \item{x}{a numeric or ordered vector}

  \item{na.rm}{ignored}
}
\value{
  the maximum value of the vector
}
\description{
  Compute the Maximum Value of a Vector
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

