\name{Median}
\alias{Median}
\title{Compute the Statistical Median of a Vector...}
\usage{Median(x, na.rm)}
\description{Compute the Statistical Median of a Vector}
\value{the median of the vector}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{x}{a vector of numeric or ordered values}
\item{na.rm}{ignored}}
