\name{median.zelig}
\alias{median.zelig}
\title{Compute the Statistical Median of a Vector...}
\usage{median.zelig(x)}
\description{Compute the Statistical Median of a Vector}
\value{the median of the vector}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{x}{a vector of numeric or ordered values}}
