\name{mi.default}
\alias{mi.default}
\title{Construct an 'mi' object from a set of data-frames...}
\usage{mi.default(obj, ..., by)}
\description{Construct an 'mi' object from a set of data-frames}
\value{an object of type `mi'}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{obj}{the first data-frame}
\item{...}{additional data-frames}
\item{by}{a character-string specifying a column of a data.frame
to use for multiple imputation}}
