\name{mi.mi}
\alias{mi.mi}
\title{Construct an 'mi' object from another 'mi' object...}
\usage{mi.mi(obj, ..., by)}
\description{Construct an 'mi' object from another 'mi' object
This cosntructor differs from an identity operation
by applying the 'by' parameter.}
\value{an object of type 'mi'}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{obj}{an object of type 'mi'}
\item{...}{ignored parameters}
\item{by}{a character-string specifying a calumn of the data.frames
used in the 'mi' object.}}
