\name{Min}
\alias{Min}
\title{Compute the Minumum Value of a Vector...}
\usage{Min(x, na.rm)}
\description{Compute the Minumum Value of a Vector}
\value{the minimum value of the vector}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{x}{a vector of numeric or ordered values}
\item{na.rm}{ignored}}
