\name{min.zelig}
\alias{min.zelig}
\title{Compute the Minumum Value of a Vector...}
\usage{min.zelig(x)}
\description{Compute the Minumum Value of a Vector}
\value{the minimum value of the vector}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{x}{a vector of numeric or ordered values}}
