\name{mode.zelig}
\alias{mode.zelig}
\title{Compute the Statistical Mode of a Vector...}
\usage{mode.zelig(x)}
\description{Compute the Statistical Mode of a Vector}
\value{the statistical mode of the vector. If two modes exist, one is
randomly selected (by design)}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{x}{a vector of numeric, factor, or ordered values}}
