\name{model.matrix.multiple}
\alias{model.matrix.multiple}
\title{Create Design Matrix of a multiple Object...}
\usage{\method{model.matrix}{multiple}(object,data,shape="compact",eqn=NULL,...)}
\description{Create Design Matrix of a \code{multiple} Object}
\details{This method is used to generate a \code{model.matrix} adhering to the
specifications in the help document "model.matrix".}
\note{This method is scheduled to be deprecated.}
\arguments{\item{object}{an object of type \code{multiple}. This represents a Zelig 3.5
formula}
\item{data}{a \code{data.frame}}
\item{shape}{a character-string specifying the shape of the matrix}
\item{eqn}{an integer specifying the number of equations}
\item{...}{ignored parameters}}
