\name{multilevel}
\alias{multilevel}
\title{Multilevel...}
\usage{multilevel(tt, data, mode, eqn, ...)}
\description{Multilevel}
\details{This function currently has no documentation, but is essential in Zelig 3.5's
implementation of formulae.}
\value{a list with the "terms" attribute specified}
\author{Kosuke Imai, Olivia Lau, Gary King and Ferdinand Alimadhi}
\arguments{\item{tt}{a terms object}
\item{data}{a \code{data.frame}}
\item{mode}{???}
\item{eqn}{an integer specifying the number of equations in a model}
\item{...}{ignored parameters}}
