\name{package.zelig}
\alias{package.zelig}
\title{Construct an Object to Describe Package Name and Version...}
\usage{package.zelig(name, version="1.0")}
\description{Construct an Object to Describe Package Name and Version}
\note{This function is deprecated.}
\value{a list containing 'name' and 'version' keys}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{name}{a character-string specifying the name of a R-package}
\item{version}{a character-string specifying the version number of the
R-package}}
