\name{param.default}
\alias{param.default}
\title{Default Method for ``param''}
\usage{
  \method{param}{default}(obj, num, ...)
}
\arguments{
  \item{obj}{ignored parameter}

  \item{num}{ignored parameter}

  \item{...}{ignored parameters}
}
\description{
  If no \code{param} function is set for a Zelig model,
  then this function will return NULL.
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

