\name{param.exp}
\alias{param.exp}
\title{Param Method for the \code{exp} Zelig Model}
\usage{
  \method{param}{exp}(obj, num, ...)
}
\arguments{
  \item{obj}{a 'zelig' object}

  \item{num}{an integer specifying the number of
  simulations to sample}

  \item{...}{ignored parameters}
}
\value{
  a list to be cast as a 'parameters' object
}
\description{
  Param Method for the \code{exp} Zelig Model
}
\note{
  This method is used by the \code{param} Zelig model
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

