\name{param.gamma}
\alias{param.gamma}
\title{param method for the `gamma' Zelig model...}
\usage{param.gamma(obj, num, ...)}
\description{param method for the `gamma' Zelig model}
\value{a list to be cast as a `parameters' object}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{obj}{a `zelig' object}
\item{num}{an integer specifying the number of simulations to sample}
\item{...}{ignored parameters}}
