\name{param.logit}
\alias{param.logit}
\title{Param Method for the \code{logit} Zelig Model}
\usage{
  param.logit(obj, num, ...)
}
\arguments{
  \item{obj}{a 'zelig' object}

  \item{num}{an integer specifying the number of
  simulations to sample}

  \item{...}{ignored parameters}
}
\value{
  a list to be cast as a 'parameters' object
}
\description{
  Param Method for the \code{logit} Zelig Model
}
\note{
  This method is used by the \code{logit} Zelig model
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

