\name{param.ls}
\alias{param.ls}
\title{Param Method for the 'ls' Zelig Model}
\usage{
  \method{param}{ls}(obj, num, \dots)
}
\arguments{
  \item{obj}{a 'zelig' object}

  \item{num}{an integer specifying the number of
  simulations to sample}

  \item{...}{ignored parameters}
}
\value{
  a list to be cast as a 'parameters' object
}
\description{
  Param Method for the 'ls' Zelig Model
}
\note{
  This method currently returns via a deprectated style
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

