\name{param.normal}
\alias{param.normal}
\title{Param Method for the 'normal' Zelig Model}
\usage{
  \method{param}{normal}(obj, num=1000, ...)
}
\arguments{
  \item{obj}{a 'zelig' object}

  \item{num}{an integer specifying the number of
  simulations to sample}

  \item{...}{ignored}
}
\value{
  a list to be cast as a 'parameters' object
}
\description{
  Param Method for the 'normal' Zelig Model
}
\note{
  This method is used by the 'normal' Zelig model
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

