\name{param.poisson}
\alias{param.poisson}
\title{Param Method for the 'poisson' Zelig Model...}
\usage{param.poisson(obj, num=1000, ...)}
\description{Param Method for the 'poisson' Zelig Model}
\note{This method is used by the 'poisson' Zelig model}
\value{a list to be cast as a 'parameters' object}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{obj}{a 'zelig' object}
\item{num}{an integer specifying the number of simulations to sample}
\item{...}{ignored}}
