\name{param}
\alias{param}
\title{Generic Method for Simulating Ancillary/Auxillary Parameters of Zelig Models...}
\usage{param(obj, num, ...)}
\description{Generic Method for Simulating Ancillary/Auxillary Parameters of Zelig Models}
\note{The 'param' function is a method meant to be overloaded by Zelig
Developers}
\value{a list to be cast as a `parameters' object}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{obj}{a `zelig' object}
\item{num}{an integer specifying the number of simulations to sample}
\item{...}{optional parameters which will likely be ignored}}
