\name{param.relogit2}
\alias{param.relogit2}
\title{Estimate Parameters for the ``relogit'' Zelig Mdoel}
\usage{
  param.relogit2(object, num, x, ...)
}
\arguments{
  \item{object}{a zelig object containing the fitted model}

  \item{num}{an integer specifying the number of
  simulations to compute}

  \item{x}{ideally we should be able to remove this
  parameter}
}
\value{
  ...
}
\description{
  Returns estimates on parameters, as well as, specifying
  link and inverse-link functions.
}

