\name{parse.formula}
\alias{parse.formula}
\title{Parse Formulas for Zelig Models}
\usage{
  parse.formula(formula, model, data = NULL)
}
\arguments{
  \item{formula}{a formula}

  \item{model}{a Zelid model}

  \item{data}{a data-frame}
}
\description{
  Parse Formulas for Zelig Models
}
\note{
  This is used typically in multinomial and multivariate
  Zelig models
}
\author{
  Kosuke Imai and Olivia Lau
}

