\name{parseFormula.list}
\alias{parseFormula.list}
\title{Parse List-Style Zelig Formulae...}
\usage{\method{parseFormula}{list}(obj, data=NULL)}
\description{Parse List-Style Zelig Formulae}
\details{This method parses a list-style Zelig formula.}
\value{an object of type "parseFormula"}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{obj}{a list of formulae}
\item{data}{a data frame}}
