\name{plot.MI.sim}
\alias{plot.MI.sim}
\title{Plot graphs of simulated multiply-imputed data}
\usage{
  \method{plot}{MI.sim}(x, ...)
}
\arguments{
  \item{x}{A zelig `sim' object, with multiply imputed data}	
  \item{...}{ignored parameters}
}
\value{
  NULL (invisibly)
}
\description{
  This function is currently unimplemented, and reserved
  for future use.
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

