\name{plot.sim}
\alias{plot.sim}
\title{Method for plotting simulations}
\usage{
  plot.sim(x, xlab = "", ...)
}
\arguments{
  \item{x}{a `sim' object}

  \item{xlab}{labels for the x-axis}

  \item{...}{parameters to be passed to the `truehist'
  function which is implicitly called for numeric
  simulations}
}
\value{
  the current graphical parameters. This is subject to
  change in future implementations of Zelig
}
\description{
  Method for plotting simulations
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

