\name{plot.simulations}
\alias{plot.simulations}
\title{Plot Any Simulation from the Zelig Core Package
Plots any simulation from the core package. In general, this function can
\emph{neatly} plot simulations containing five of the popular ``quantities
of interest'' - ``Expected Values: E(Y|X)'', ``Predicted Values: Y|X'',
``Expected Values (for X1): E(Y|X1)'', ``Predicted Values (for X1): Y|X1''
and ``First Differences: E(Y|X1) - E(Y|X)''.}
\usage{
  plot.simulations(x, ...)
}
\arguments{
  \item{x}{an object}

  \item{....}{parameters passed to the ``plot'' and
  ``barplot'' functions}
}
\value{
  the original graphical parameters
}
\description{
  Plot Any Simulation from the Zelig Core Package Plots any
  simulation from the core package. In general, this
  function can \emph{neatly} plot simulations containing
  five of the popular ``quantities of interest'' -
  ``Expected Values: E(Y|X)'', ``Predicted Values: Y|X'',
  ``Expected Values (for X1): E(Y|X1)'', ``Predicted Values
  (for X1): Y|X1'' and ``First Differences: E(Y|X1) -
  E(Y|X)''.
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

