\name{print.qi}
\alias{print.qi}
\title{Print a Quantity of Interest in Human-Readable Form}
\usage{
  print.qi(x, ...)
}
\arguments{
  \item{x}{a qi object}

  \item{...}{ignored parameters}
}
\value{
  the object that was printed (invisibly)
}
\description{
  Print a Quantity of Interest in Human-Readable Form
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

