\name{print.qi.summarized}
\alias{print.qi.summarized}
\title{Print Method for Summarized Quantities of Interest}
\usage{
  \method{print}{qi.summarized}(x, \dots)
}
\arguments{
  \item{x}{a 'summarized.qi' object}

  \item{...}{parameters to be passed to the specific print
  functions}
}
\value{
  x (invisibly)
}
\description{
  Print Method for Summarized Quantities of Interest
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}
\seealso{
  \link{special_print_MATRIX} and \link{special_print_LIST}
}

