\name{print.setx.mi}
\alias{print.setx.mi}
\title{Print values of multiply-imputed `setx' objects...}
\usage{print.setx.mi(x, ...)}
\description{Print values of multiply-imputed `setx' objects}
\value{NULL}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{x}{a `setx.mi' object}
\item{...}{ignored parameters}}
