\name{print.setx}
\alias{print.setx}
\title{Print values of `setx' objects}
\usage{
  print.setx(x, ...)
}
\arguments{
  \item{x}{a `setx' object}

  \item{...}{ignored parameters}
}
\value{
  the value of x (invisibly)
}
\description{
  Print values of `setx' objects
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

