\name{print.sim}
\alias{print.sim}
\title{Print values of `sim' objects
This function is currently unimplemented, and
included for future development}
\usage{
  print.sim(x, ...)
}
\arguments{
  \item{x}{a `sim' object (ignored)}

  \item{...}{ignored parameters}
}
\value{
  NULL (invisibly)
}
\description{
  Print values of `sim' objects This function is currently
  unimplemented, and included for future development
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

