\name{print.summary.MCMCZelig}
\alias{print.summary.MCMCZelig}
\title{Print a Summary MCMCZelig Object}
\usage{
  print summary.MCMCZelig
}
\arguments{
  \item{x}{an "MCMCZelig" object}

  \item{digits}{a numeric specifying the precision of the
  summary object}

  \item{...}{ignored parameters}
}
\value{
  a \code{summary.MCMCZelig} object
}
\description{
  This method prints a summary object for \code{MCMCZelig}
  objects
}

