\name{print.summary.pooled.sim}
\alias{print.summary.pooled.sim}
\title{Print a Summary of a Set of Pooled Simulated Interests}
\usage{
  print.summary.pooled.sim(x, ...)
}
\arguments{
  \item{x}{a ``summary.pooled.sim'' object, containing
  summarized information about simulated quantities of
  interest}

  \item{...}{1}
}
\value{
  a ``summary.pooled.sim'' object storing the replicated
  quantities of interest
}
\description{
  Returns a
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

