\name{print.summarySim.MI}
\alias{print.summarySim.MI}
\title{Print Multiply Imputed Simulations Summary}
\usage{
  \method{print}{summarySim.MI}(x, digits=3, ...)
}
\arguments{
  \item{x}{a 'summarySim.MI' object}

  \item{digits}{an integer specifying the number of digits
  of precision to print}

  \item{...}{ignored parameters}
}
\description{
  Prints summary information about Multiply Imputed Fits
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

