\name{print.zelig}
\alias{print.zelig}
\title{Print values of `zelig' objects...}
\usage{print.zelig(x, ...)}
\description{Print values of `zelig' objects}
\value{the `zelig' object (invisibly)}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{x}{a `zelig' object}
\item{...}{ignored parameters}}
