\name{reduceMI}
\alias{reduceMI}
\title{Reduce MI Formulas...}
\usage{reduceMI(f)}
\description{Reduce MI Formulas
Take a formula in any of the reduced form or in a structural form and return
the most reduced form of that formula}
\note{This formula is used primarily by 'zelig2' functions of multivariate
Zelig models}
\author{Ferdinand Alimadhi, Kosuke Imai, and Olivia Lau}
\arguments{\item{f}{a formula}}
