\name{register.default}
\alias{register.default}
\title{Construct a list of available generic methods available for the...}
\usage{register.default(obj)}
\description{Construct a list of available generic methods available for the
object used to fit the Zelig model}
\value{a character vector containing names of S3 methods for the statistical model}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{obj}{a `zelig' object}}
