\name{repl}
\alias{repl}
\title{Generic Method for Replicating Data...}
\usage{repl(object, ...)}
\description{Generic Method for Replicating Data}
\value{a replicated object}
\author{Kosuke Imai and Olivia Lau \email{mowen@iq.harvard.edu}}
\arguments{\item{object}{a 'zelig' object}
\item{...}{parameters}}
