\name{replace.call}
\alias{replace.call}
\title{Hook to Update the Zelig Call with the Appropriate Call Object...}
\usage{replace.call(zobj, call1, call2)}
\description{Hook to Update the Zelig Call with the Appropriate Call Object}
\note{This function is used internally by Zelig, and currently deprecated.}
\value{the 'zelig' object with a modified 'call' slot}
\arguments{\item{zobj}{a 'zelig' object}
\item{call1}{the original call to Zelig}
\item{call2}{the manuafactured call to the model fitting function}}
