\name{reset.mi}
\alias{reset.mi}
\title{Reset method for mi objects...}
\usage{\method{reset}{mi}(obj, ...)}
\description{Reset method for \code{mi} objects}
\value{the same object with the iterator reset}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{obj}{an \code{mi} object}
\item{...}{ignored parameters}}
