\name{Reset}
\alias{Reset}
\title{Reset Counter for \code{mi} Object}
\usage{
  Reset(obj)
}
\arguments{
  \item{obj}{an \code{mi} object}
}
\value{
  the value 1 (invisibly)
}
\description{
  This function sets the next frame of the iterator to be
  its first value.
}
\note{
  This method is used for its side-effect. Its return-value
  is of almost no significance.
}

