\name{retrieve.factors}
\alias{retrieve.factors}
\title{Retrieve a list of all the factors in the specified column...}
\usage{retrieve.factors(frames, by)}
\description{Retrieve a list of all the factors in the specified column}
\value{a list with entries corresponding to the value of the \code{by}
parameter. The value of each index is unique values of the
\code{data.frame}'s columns.}
\arguments{\item{frames}{a list of \code{data.frame} objects}
\item{by}{a character-vector specifying columns to search for factors within}}
