\name{rocplot}
\alias{rocplot}
\title{rocplot...}
\usage{rocplot(y1, y2, fitted1, fitted2, cutoff=seq(from = 0, to = 1, length =
    100), lty1="solid", lty2="dashed", lwd1=par("lwd"),
    lwd2=par("lwd"), col1=par("col"), col2=par("col"), main="ROC
    Curve", xlab="Proportion of 1's Correctly Predicted",
    ylab="Proportion of 0's Correctly Predicted", plot=TRUE, ...)}
\description{rocplot}
\value{a list of relevant statistics}
\arguments{\item{y1}{}
\item{y2}{}
\item{fitted1}{}
\item{fitted2}{}
\item{cutoff}{}
\item{lty1}{}
\item{lty2}{}
\item{lwd1}{}
\item{lwd2}{}
\item{col1}{}
\item{col2}{}
\item{main}{}
\item{xlab}{}
\item{ylab}{}
\item{plot}{}
\item{...}{}}
