\name{setx}
\alias{setx}
\title{Generic Function for Setting Explanatory Variables and Counterfactuals...}
\usage{setx(obj, ..., data)}
\description{Generic Function for Setting Explanatory Variables and Counterfactuals}
\value{a setx object}
\author{Matt Owen and Olivia Lau and Kosuke Imai \email{mowen@iq.harvard.edu}}
\arguments{\item{obj}{}
\item{...}{}
\item{data}{}}
