\name{special_print_LIST}
\alias{.print.qi.summarized.LIST}
\alias{special_print_LIST}
\title{Method for Printing Summarized QI's in a List Form}
\usage{
  .print.qi.summarized.LIST(x, ...)
}
\arguments{
  \item{x}{a 'summarized.qi' object}

  \item{...}{additional parameters to be used by the
  'print.matrix' method}
}
\value{
  x (invisibly)
}
\description{
  Method for Printing Summarized QI's in a List Form
}
\note{
  This function is used internall by Zelig
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

