\name{special_print_MATRIX}
\alias{.print.qi.summarized.MATRIX}
\alias{special_print_MATRIX}
\title{Method for Printing Summarized QI's in a Matrix Form}
\usage{
  .print.qi.summarized.MATRIX(x, ...)
}
\arguments{
  \item{x}{a 'summarized.qi' object}

  \item{...}{additional parameters}
}
\value{
  x (invisibly)
}
\description{
  Method for Printing Summarized QI's in a Matrix Form
}
\note{
  This function is used internall by Zelig
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

