\name{structuralToReduced}
\alias{structuralToReduced}
\title{Transform the Multilevel's Structural Formulas Into Reduced Form}
\usage{
  structuralToReduced(f)
}
\arguments{
  \item{f}{a list of formulas}
}
\value{
  a formula in reduced form
}
\description{
  Transform the Multilevel's Structural Formulas Into
  Reduced Form
}
\author{
  Ferdinand Alimadhi, Kosuke Imai, and Olivia Lau
}

