\name{summarize.default}
\alias{summarize.default}
\title{Summarize Simualted Quantities of Interest...}
\usage{\method{summarize}{default}(q)}
\description{Summarize Simualted Quantities of Interest}
\value{a 'summarized.qi' object}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{q}{a 'qi' object, storing simulations of quantities of interest}}
