\name{summarize}
\alias{summarize}
\title{Generic methonf for summarizing simualted quantities of interest...}
\usage{summarize(obj)}
\description{Generic methonf for summarizing simualted quantities of interest}
\value{a \code{summarized.qi} object}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{obj}{a \code{qi} object, storing simulations of quantities of interest}}
