\name{summary.MI.sim}
\alias{summary.MI.sim}
\title{Method for summarizing simulations of multiply imputed quantities of interest...}
\usage{summary.MI.sim(object, ...)}
\description{Method for summarizing simulations of multiply imputed quantities of interest}
\value{a `summarized.MI.sim' object}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{object}{a `MI.sim' object}
\item{...}{ignored parameters}}
