\name{summary.pooled.sim}
\alias{summary.pooled.sim}
\title{Return a Summary of a Set of Pooled Simulated Interests}
\usage{
  summary.pooled.sim(object, ...)
}
\arguments{
  \item{object}{a ``pooled.sim'' object, containing
  information about simulated quantities of interest}

  \item{...}{1}
}
\value{
  a ``summary.pooled.sim'' object storing the replicated
  quantities of interest
}
\description{
  Returns a
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

