\name{summary.Relogit}
\alias{summary.Relogit}
\title{Summary for ``Relogit'' Fitted Model}
\usage{
  \method{summary}{Relogit}(object, ...)
}
\arguments{
  \item{object}{a ``Relogit'' object}

  \item{...}{other parameters}
}
\value{
  a ``summary.relogit'' object
}
\description{
  Summarize important components of the ``relogit'' model
}

