\name{summary.relogit.sim}
\alias{summary.relogit.sim}
\title{Summary for ``relogit'' Fitted Model}
\usage{
  summary.relogit.sim(object, ...)
}
\arguments{
  \item{object}{a ``relogit'' object}

  \item{...}{other parameters}
}
\value{
  a ``summary.relogit'' object
}
\description{
  Summarize important components of the ``relogit'' model
}

