\name{summary.Relogit2}
\alias{summary.Relogit2}
\title{Summary for ``Relogit2'' Fitted Model}
\usage{
  summary.Relogit2(object, ...)
}
\arguments{
  \item{object}{a ``Relogit2'' object}

  \item{...}{other parameters}
}
\value{
  a ``summary.relogit2'' object
}
\description{
  Summarize important components of the ``relogit'' model
}

