\name{summary.sim}
\alias{summary.sim}
\title{Method for summarizing simulations of quantities of interest}
\usage{
  \method{summary}{sim}(object, ...)
}
\arguments{
  \item{object}{a 'MI.sim' object}

  \item{...}{ignored parameters}
}
\value{
  a 'summarized.MI.sim' object
}
\description{
  Return a ``summary.sim'' object (typically for display)
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

