\name{table.levels}
\alias{table.levels}
\title{Create a table, but ensure that the correct...}
\usage{table.levels(x, levels, ...)}
\description{Create a table, but ensure that the correct
columns exist. In particular, this allows for
entires with zero as a value, which is not
the default for standard tables}
\value{a table}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{x}{a vector}
\item{levels}{a vector of levels}
\item{...}{parameters for table}}
