\name{terms.multiple}
\alias{terms.multiple}
\title{Extract Terms from a \code{multiple} Object}
\usage{
  \method{terms}{multiple}(x, data=NULL,...)
}
\arguments{
  \item{x}{a Zelig v3.5 formula}

  \item{data}{a \code{data.frame}}

  \item{...}{ignored parameters}
}
\description{
  Extracts terms from Zelig-3.5-style formulae. This
  function is scheduled for removal.
}
\author{
  Kosuke Imai, Olivia Lau, Gary King and Ferdinand Alimadhi
}

