\name{terms.vglm}
\alias{terms.vglm}
\title{Model Terms for 'vglm' Models...}
\usage{\method{terms}{vglm}(x, ...)}
\description{Model Terms for 'vglm' Models}
\value{the models terms of this fitted model object}
\author{Ferdinand Alimadhi, Kosuke Imai and Olivia Lau}
\arguments{\item{x}{a fitted model object from the VGAM library}
\item{...}{ignored parameters}}
