\name{terms.zelig}
\alias{terms.zelig}
\title{Model Terms for a Zelig Object...}
\usage{\method{terms}{zelig}(x, ...)}
\description{Model Terms for a Zelig Object}
\details{This method simply extracts the model terms for the fitted model passed to 
the \code{zelig} function.}
\value{terms of the original fitted model}
\arguments{\item{x}{a \code{zelig} object}
\item{...}{forwarded parameters}}
