\name{TexCite}
\alias{TexCite}
\title{Get a TeX-style Citation...}
\usage{TexCite(model)}
\description{Get a TeX-style Citation}
\value{a string to be rendered as part of a LaTeX-style document}
\arguments{\item{model}{a character-string specifying the name of the Zelig model of which 
to describe in TeX-style}}
