\name{toBuildFormula}
\alias{toBuildFormula}
\title{Build Formula ???}
\usage{
  toBuildFormula(Xnames, sepp = "+")
}
\arguments{
  \item{Xnames}{a character-vector}

  \item{sepp}{a seperator (???)}
}
\value{
  a character-string
}
\description{
  This function builds a formula
}
\author{
  ???
}

