\name{toString.call}
\alias{toString.call}
\title{Convert \code{call} Object to a String}
\usage{
  \method{toString}{call}(x, ...)
}
\arguments{
  \item{x}{a \code{call} object}

  \item{...}{ignored parameters}
}
\value{
  a character-string representing the \code{call} object
}
\description{
  This method concerts \code{call} objects into a simple,
  intuitive human-readable form.
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

