\name{ucfirst}
\alias{ucfirst}
\title{Uppercase First Letter of a String}
\usage{
  ucfirst(str)
}
\arguments{
  \item{str}{a vector of charaqcter-strings}
}
\value{
  a vector of character strings
}
\description{
  This method sets the first character of a string to its
  uppercase, sets all other characters to lowercase.
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

