\name{user.prompt}
\alias{user.prompt}
\title{Prompt User}
\usage{user.prompt(msg)}
\description{Prompts user to hit enter}
\value{This function is used for its side effects}
\note{This function is primarily used by Zelig demo scripts}
\arguments{\item{msg}{a character-string, specifying a message to be displayed}}
