\name{vcov.glm.robust}
\alias{vcov.glm.robust}
\title{Return Variance-Covariance Matrix}
\usage{
  vcov.glm.robust(object, ...)
}
\arguments{
  \item{object}{}

  \item{...}{}
}
\value{
  a NxN matrix, where N is the number of parameters in the
  fitted model.
}
\description{
  Returns the variance-covariance matrix produced for this
  model.
}

