\docType{package}
\name{Zelig-package}
\alias{Zelig}
\alias{Zelig-package}
\title{Zelig Everyone's Statistical Software}
\description{
  Zelig is an easy-to-use program that can estimate, and
  help interpret the results of, an enormous range of
  statistical models. It literally is ``everyone's
  statistical software'' because Zelig's simple unified
  framework incorporates everyone else's (R) code. We also
  hope it will become ``everyone's statistical software''
  for applications and teaching, and so have designed Zelig
  so that anyone can easily use it or add their programs to
  it.  Zelig also comes with infrastructure that
  facilitates the use of any existing method, such as by
  allowing multiply imputed data for any model, and
  mimicking the program Clarify (for Stata) that takes the
  raw output of existing statistical procedures and
  translates them into quantities of direct interest.
}
\details{
  \tabular{ll}{ Package: \tab Zelig\cr Version: \tab
  4.0-11\cr Date: \tab 2011-08-15\cr Depends: \tab R (>=
  2.8.1), MASS, survival, methods Suggests: mvtnorm,
  MCMCpack\cr License: \tab GPL version 2 or newer\cr URL:
  \tab http://gking.harvard.edu/zelig\cr }
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}, Kosuke Imai,
  Olivia Lau, and Gary King Maintainer: Matt Owen
  \email{monwe@iq.harvard.edu}
}
\seealso{
  zelig setx sim
}
\keyword{package}

