\name{zelig.kin}
\alias{zelig.kin}
\title{Create a Zelig Object from an 'MI' Object...}
\usage{zelig.kin(parent, fitted, data)}
\description{Create a Zelig Object from an 'MI' Object}
\note{This function is exclusively used internally by Zelig}
\value{a zelig object}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{parent}{the parent model (MI)}
\item{fitted}{the fitted model object}
\item{data}{a new data-frame}}
