\name{zelig2factor.bayes}
\alias{zelig2factor.bayes}
\title{Interface between the Zelig Model factor.bayes and
the Pre-existing Model-fitting Method}
\usage{
  zelig2factor.bayes(formula, factors = 2, burnin = 1000,
    mcmc = 20000, verbose = 0, ..., data)
}
\arguments{
  \item{formula}{a formula}

  \item{...}{additonal parameters}

  \item{data}{a data.frame}
}
\value{
  a list specifying '.function'
}
\description{
  Interface between the Zelig Model factor.bayes and the
  Pre-existing Model-fitting Method
}

