\name{zelig2lognorm}
\alias{zelig2lognorm}
\title{Interface between the Zelig Model lognorm and
the Pre-existing Model-fitting Method}
\usage{
  zelig2lognorm(formula, ..., robust = FALSE,
    cluster = NULL, data)
}
\arguments{
  \item{formula}{a formula}

  \item{...}{additonal parameters}

  \item{data}{a data.frame}
}
\value{
  a list specifying '.function'
}
\description{
  Interface between the Zelig Model lognorm and the
  Pre-existing Model-fitting Method
}

