\name{zelig2ls}
\alias{zelig2ls}
\title{Interface between ls model and Zelig...}
\usage{zelig2ls(formula, ..., data, weights)}
\description{Interface between ls model and Zelig
This function is exclusively for use by the `zelig' function}
\value{a list to be coerced into a zelig.call object}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{formula}{a formula}
\item{weights}{a numeric vector}
\item{...}{ignored parameters}
\item{data}{a data.frame}}
