\name{zelig2normal.bayes}
\alias{zelig2normal.bayes}
\title{Interface between the Zelig Model normal.bayes and
the Pre-existing Model-fitting Method}
\usage{
  zelig2normal.bayes(formula, burnin = 1000, mcmc = 10000,
    verbose = 0, ..., data)
}
\arguments{
  \item{formula}{a formula}

  \item{...}{additonal parameters}

  \item{data}{a data.frame}
}
\value{
  a list specifying '.function'
}
\description{
  Interface between the Zelig Model normal.bayes and the
  Pre-existing Model-fitting Method
}

