\name{zelig2relogit}
\alias{zelig2relogit}
\title{Zelig2 bridge function}
\usage{
  zelig2relogit(formula, ..., tau = NULL,
    bias.correct = NULL, case.control = NULL, data)
}
\arguments{
  \item{formula}{a formula object}

  \item{model}{a chacter}

  \item{...}{}

  \item{bias.correct}{...}

  \item{case.control}{...}

  \item{data}{a data.frame}
}
\value{
  ...
}
\description{
  ...
}

