\name{zelig2threesls}
\alias{zelig2threesls}
\title{Interface between the Zelig Model threesls and
the Pre-existing Model-fitting Method}
\usage{
  zelig2threesls(formula, ..., data)
}
\arguments{
  \item{formula}{a formula}

  \item{...}{additonal parameters}

  \item{data}{a data.frame}
}
\value{
  a list specifying '.function'
}
\description{
  Interface between the Zelig Model threesls and the
  Pre-existing Model-fitting Method
}

