\name{ZeligDescribeModel}
\alias{ZeligDescribeModel}
\title{Produce a 'description' Object from the Name of a Model}
\usage{
  ZeligDescribeModel(model.name)
}
\arguments{
  \item{model.name}{a character-string specifying a Zelig
  model}
}
\value{
  a 'description' object specified by the 'model.name'
  parameter. This object is created by executing the
  specified Zelig models' 'describe' function
}
\description{
  Produce a 'description' Object from the Name of a Model
}
\note{
  The 'description' object is a list-style object
  containing citation information
}

