\name{.copy.templates}
\alias{.copy.templates}
\title{Search for, Copy, and Customize Template for a Newly Create Zelig Package...}
\usage{.copy.templates(model, pkg, path)}
\description{Search for, Copy, and Customize Template for a Newly Create Zelig Package
This is used internally by \code{zelig.skeleton}}
\value{This function is used for its side-effects.}
\arguments{\item{model}{a character-string specifying the name of the model}
\item{the}{file template to search for and copy}
\item{pkg}{a character-string specifying the name of the package}
\item{path}{a character-string specifying the base path to the package's u
parent directory}}
