\name{alpha}
\alias{alpha}
\title{Extract ancillary parameters from
`parameters' objects}
\usage{
  alpha(param)
}
\arguments{
  \item{param}{a `parameters' object}
}
\value{
  the ancillary parameters \emph{specified} for the
  statistical model
}
\description{
  Extract ancillary parameters from `parameters' objects
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

