\name{as.description.description}
\alias{as.description.description}
\title{description -> description}
\usage{
  \method{as.description}{description}(descr, ...)
}
\arguments{
  \item{descr}{an object of type 'description'}

  \item{...}{ignored}
}
\value{
  the same object
}
\description{
  Identity operation on a description object.
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

