\name{as.description.description}
\alias{as.description.description}
\title{description -> description...}
\usage{as.description.description(descr, ...)}
\description{description -> description}
\value{the same object}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{descr}{an object of type 'description'}
\item{...}{ignored}}
