\name{as.description.list}
\alias{as.description.list}
\title{list -> description...}
\usage{as.description.list(descr, ...)}
\description{list -> description}
\value{an object of type 'description'}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{descr}{a list}
\item{...}{ignored}}
