\name{as.description.list}
\alias{as.description.list}
\title{list -> description}
\usage{
  as.description.list(descr, ...)
}
\arguments{
  \item{descr}{a list}

  \item{...}{ignored}
}
\value{
  an object of type 'description'
}
\description{
  list -> description
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

