\name{as.parameters.default}
\alias{as.parameters.default}
\title{?}
\usage{as.parameters.default(params, num, ...)}
\description{??? -> parameters}
\note{This function should be deprecated.}
\value{the object passed in}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{params}{any non-supported data-type}
\item{num}{an integer specifying the number of simulations to compute}
\item{...}{ignored}}
