\name{as.parameters.parameters}
\alias{as.parameters.parameters}
\title{parameters -> parameters...}
\usage{as.parameters.parameters(params, ...)}
\description{parameters -> parameters
This is merely an identity function when casting 'parameters' objects into
'parameters'.}
\value{the same parameter object}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{params}{a parameters object}
\item{...}{ignored parameters}}
