\name{as.summarized.list}
\alias{as.summarized.list}
\title{list -> summarized.qi
Convert a list into a ``summarized.qi'' object}
\usage{
  \method{as.summarized}{list}(x, ...)
}
\arguments{
  \item{x}{a list}

  \item{...}{ignored parameters}
}
\value{
  a ``summarized.qi'' object
}
\description{
  list -> summarized.qi Convert a list into a
  ``summarized.qi'' object
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

