\name{as.summarized}
\alias{as.summarized}
\title{Generic Method for Casting Objectst as 'summarized' Objects...}
\usage{as.summarized(x, ...)}
\description{Generic Method for Casting Objectst as 'summarized' Objects
This function is particularly for use by the 'summarize' method, which
summarizes the simulations taken from the 'qi' method.
The generic function 'summary' when applied to a Zelig Simulation implicitly
uses this function.}
\note{This is made available on the Global namespace as a matter of
potential future compliancy.}
\value{a 'summarized.qi' object}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{x}{an object}
\item{...}{unspecified parameters}}
