\name{as.summarized.summarized.qi}
\alias{as.summarized.summarized.qi}
\title{summarized.qi -> summarized.qi}
\usage{
  \method{as.summarized}{summarized.qi}(x, ...)
}
\arguments{
  \item{x}{an object of type 'summarized.qi'}

  \item{...}{ignored parameters}
}
\value{
  the same 'summarized.qi' object
}
\description{
  Identity operation on ``summarized.qi'' objects
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

