\name{as.summarized.summarized.qi}
\alias{as.summarized.summarized.qi}
\title{summarized.qi -> summarized.qi
This is an identity operation}
\usage{
  as.summarized.summarized.qi(x, ...)
}
\arguments{
  \item{x}{an object of type 'summarized.qi'}

  \item{...}{ignored parameters}
}
\value{
  the same 'summarized.qi' object
}
\description{
  summarized.qi -> summarized.qi This is an identity
  operation
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

