\name{cite}
\alias{cite}
\title{Citation information for a 'description' object...}
\usage{cite(descr)}
\description{Citation information for a 'description' object}
\value{a character-string giving citation info}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{descr}{an object of type 'description'}}
