\name{constructDesignMatrix}
\alias{constructDesignMatrix}
\title{Construct Design Matrix from
Construct and return a design matrix based on a tiny data-frame (single-row).}
\usage{
  constructDesignMatrix(data, formula)
}
\arguments{
  \item{data}{...}

  \item{formula}{...}
}
\value{
  ...
}
\description{
  Construct Design Matrix from Construct and return a
  design matrix based on a tiny data-frame (single-row).
}

