\name{describe.default}
\alias{describe.default}
\title{Default describe function for an arbitrary model...}
\usage{describe.default(...)}
\description{Default describe function for an arbitrary model}
\details{This method exists solely as a backup when an author does not
contribute a `describe' function for their model}
\value{a list to be processed by `as.description'}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{...}{dummy parameters purely to cast the correct object.
That is, the parameters of the function should not
be referenced specifically}}
