\name{describe.gamma}
\alias{describe.gamma}
\title{Describe the gamma model to Zelig...}
\usage{describe.gamma(...)}
\description{Describe the \code{gamma} model to Zelig}
\value{a list of important information}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{...}{ignored parameters}}
