\name{describe.logit}
\alias{describe.logit}
\title{Describe a `logit' model to Zelig}
\usage{
  \method{describe}{logit}(...)
}
\arguments{
  \item{...}{ignored parameters}
}
\value{
  a list to be processed by `as.description'
}
\description{
  Describe a `logit' model to Zelig
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

