\name{describe.ls}
\alias{describe.ls}
\title{Describe a \code{ls} model to Zelig}
\usage{
  \method{describe}{ls}(...)
}
\arguments{
  \item{...}{ignored parameters}
}
\value{
  a list to be processed by \code{as.description}
}
\description{
  Describe a \code{ls} model to Zelig
}
\note{
  \code{ls} stands for "least squares fit"
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

