\name{describe.ls}
\alias{describe.ls}
\title{Describe a ls model to Zelig...}
\usage{describe.ls(...)}
\description{Describe a \code{ls} model to Zelig}
\note{\code{ls} stands for "least squares fit"}
\value{a list to be processed by \code{as.description}}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{...}{ignored parameters}}
