\name{describe.negbinom}
\alias{describe.negbinom}
\title{Describe the negbinom model to Zelig...}
\usage{describe.negbinom(...)}
\description{Describe the \code{negbinom} model to Zelig}
\note{\code{negbinom} stands for "negative binomial"}
\value{a list to be processed by \code{as.description}}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{...}{ignored parameters}}
