\name{describe.negbinom}
\alias{describe.negbinom}
\title{Describe the \code{negbinom} model to Zelig}
\usage{
  \method{describe}{negbinom}(...)
}
\arguments{
  \item{...}{ignored parameters}
}
\value{
  a list to be processed by \code{as.description}
}
\description{
  Describe the \code{negbinom} model to Zelig
}
\note{
  \code{negbinom} stands for "negative binomial"
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

