\name{describe.normal}
\alias{describe.normal}
\title{Describe the \code{normal} model to Zelig}
\usage{
  describe.normal(...)
}
\arguments{
  \item{...}{ignored parameters}
}
\value{
  a list to be processed by `as.description'
}
\description{
  Describe the \code{normal} model to Zelig
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

