\name{describe.poisson}
\alias{describe.poisson}
\title{Describe the `poisson' model to Zelig}
\usage{
  \method{describe}{poisson}(...)
}
\arguments{
  \item{...}{ignored parameters}
}
\value{
  a list to be processed by `as.description'
}
\description{
  Describe the `poisson' model to Zelig
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

