\name{describe.probit}
\alias{describe.probit}
\title{Describe the `probit' model to Zelig}
\usage{
  describe.probit(...)
}
\arguments{
  \item{...}{ignored parameters}
}
\value{
  a list to be processed by `as.description'
}
\description{
  Describe the `probit' model to Zelig
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

