\name{describe}
\alias{describe}
\title{Method to describe a model to Zelig}
\usage{
  describe(...)
}
\arguments{
  \item{...}{parameters which are typically ignored}
}
\value{
  a list to be processed by `as.description'
}
\description{
  Method to describe a model to Zelig
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

