\name{eval.in}
\alias{eval.in}
\title{Attach Environment and Evaluate Expression}
\usage{
  eval.in(...)
}
\arguments{
  \item{...}{Two parameters: The first is an expression to
  evaluate, and the second the environment to evaluate the
  expression in}
}
\value{
  evaluation of ``..1'' within the environment ``..2''
}
\description{
  Run in an insulated environment
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

