\name{eval.in}
\alias{eval.in}
\title{Attach Environment and Evaluate Expression}
\usage{
  eval.in(...)
}
\arguments{
  \item{..1}{an expression to evaluate}

  \item{..2}{the environment to evaluate the expression in}
}
\value{
  evaluation of ..1
}
\description{
  Attach Environment and Evaluate Expression
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

