\name{GetObject}
\alias{GetObject}
\title{Extract the fitted model object from the Zelig object}
\usage{
  GetObject(obj)
}
\arguments{
  \item{obj}{an object of type `zelig'}
}
\value{
  the fitted model object
}
\description{
  Extract the fitted model object from the Zelig object
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

