\name{getPredictorTerms}
\alias{getPredictorTerms}
\title{Get Predictor Terms from Zelig-style Formulae...}
\usage{getPredictorTerms(x, ...)}
\description{Get Predictor Terms from Zelig-style Formulae}
\details{This function extracts the predictor terms from a Zelig-style object.}
\note{This function is used exclusively in the development of Zelig-core.}
\value{a character-vector or NA}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{x}{a Zelig-style formula ('formula' or 'list')}
\item{...}{ignored parameters}}
