\name{getResponseTerms.list}
\alias{getResponseTerms.list}
\title{Get Response Terms from a List-style Formula}
\usage{
  \method{getResponseTerms}{list}(x, ...)
}
\arguments{
  \item{x}{a list of formulae}

  \item{...}{ignored parameters}
}
\value{
  a character-vector specifying the response terms of the
  formula
}
\description{
  This method gets the response terms from a standard
  formula
}
\author{
  Matt Owen
}

