\name{getResponseTerms}
\alias{getResponseTerms}
\title{Get Response Terms from a Zelig-style Formula...}
\usage{getResponseTerms(x, ...)}
\description{Get Response Terms from a Zelig-style Formula}
\details{This method acquires the response variables from Zelig-style input.}
\value{a character-vector specifying a the of response terms in this formula}
\arguments{\item{x}{a formula or list of formulae}
\item{...}{ignored parameters}}
