\name{help.zelig}

\alias{help.zelig}

\title{HTML Help for Zelig Commands and Models}

\description{
  The \code{help.zelig} command launches html help for Zelig commands
  and supported models.  The full manual is available online at
  \url{http://gking.harvard.edu/zelig}.
  }

\usage{
help.zelig(...)
}

\arguments{
  \item{...}{a Zelig command or model. 
    \code{help.zelig(command)} will take you to an index of Zelig
    commands and \code{help.zelig(model)} will take you to a list of
    models. }
  }

\seealso{The complete document is available online at
  \url{http://gking.harvard.edu/zelig}.  
}

\author{
  Kosuke Imai <\email{kimai@princeton.edu}>; Gary King
  <\email{king@harvard.edu}>; Olivia Lau<\email{olau@fas.harvard.edu}>
}

\keyword{documentation}








