\name{ignore}
\alias{ignore}
\title{Constructor for the 'ignore' class...}
\usage{ignore(default, type="no pass")}
\description{Constructor for the 'ignore' class
This class is included for future use, and is currently
not used in any Zelig model. It is designed for use with
zelig2* functions}
\value{an 'ignore' object}
\author{Matt Owen \emph{mowen@iq.harvard.edu}}
\arguments{\item{default}{default value}
\item{type}{ignored parameter}}
