\name{is.qi}
\alias{is.qi}
\title{Test If Value is Interpretable as a QI}
\usage{
  is.qi(qi)
}
\arguments{
  \item{qi}{a potential quantity of interest}
}
\value{
  a logical specifying whether this value should or
  should-not be output
}
\description{
  Test If Value is Interpretable as a QI
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

