\name{is.valid.qi.list}
\alias{is.valid.qi.list}
\title{Check If Object Is a List of Valid Quantities of Interest...}
\usage{is.valid.qi.list(x)}
\description{Check If Object Is a List of Valid Quantities of Interest}
\value{TRUE or FALSE}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{x}{an object to be tested}}
