\name{is.zelig.compliant}
\alias{is.zelig.compliant}
\title{Whether a R-Package Contains a 'Yes' in its DESCRIPTION File's 'Zelig' Field...}
\usage{is.zelig.compliant(package="")}
\description{Whether a R-Package Contains a 'Yes' in its DESCRIPTION File's 'Zelig' Field}
\note{This package was used internally to determine whether an R-package is
Zelig compliant, but is now likely deprecated.}
\value{whether the package's DESCRIPTION file specifies Zelig-compliancy}
\seealso{is.zelig.package}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{package}{a character-string specifying an installed R-package}}
