\name{is.zelig.package}
\alias{is.zelig.package}
\title{Wether an Installed R-Pack Depends on Zelig}
\usage{
  is.zelig.package(package = "")
}
\arguments{
  \item{package}{a character-string naming a package}
}
\value{
  whether this package depends on Zelig
}
\description{
  Wether an Installed R-Pack Depends on Zelig
}
\note{
  This package was used internally to determine whether an
  R-package is Zelig compliant, but is now likely
  deprecated. This test is useless if not paired with
}

