\name{length.mi}
\alias{length.mi}
\title{Length Method for 'mi' Objects...}
\usage{\method{length}{mi}(x)}
\description{Length Method for 'mi' Objects
Compute the length of the number data-frames to be expected.}
\note{This function is primarily used internally by Zelig.}
\note{This function is exported, since it could be a conceivably useful
feature for end-users.}
\value{an integer specifying the number of subsetted data-frame's that are
contained within within the 'mi' object}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{x}{an object of type 'mi'}}
