\name{link}
\alias{link}
\title{Method for extracting the link function from 'parameters' objects}
\usage{
  link(param)
}
\arguments{
  \item{param}{a 'parameters' object}
}
\value{
  the link function specified by the `param' function for
  the given Zelig model
}
\description{
  Method for extracting the link function from 'parameters'
  objects
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

