\name{linkinv}
\alias{linkinv}
\title{Method for extracting the inverse link function from 'parameters' objects}
\usage{
  linkinv(param)
}
\arguments{
  \item{param}{a 'parameters' object}
}
\value{
  the inverse link function specified by the 'param'
  function for the given Zelig model
}
\description{
  Method for extracting the inverse link function from
  'parameters' objects
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

