\name{Max}
\alias{Max}
\title{Compute the Maximum Value of a Vector...}
\usage{Max(x, na.rm)}
\description{Compute the Maximum Value of a Vector}
\value{the maximum value of the vector}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{x}{a numeric or ordered vector}
\item{na.rm}{ignored}}
