\name{mi}
\alias{mi}
\title{Generic constructor for `mi' objects...}
\usage{mi(..., by)}
\description{Generic constructor for `mi' objects}
\value{an object of type `mi'}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{...}{an object or set objects to cast as an `mi' object}
\item{by}{a character-string specifying a column name
in a data.frame to subset}}
