\name{Min}
\alias{Min}
\title{Compute the Minumum Value of a Vector}
\usage{
  Min(x, na.rm = NULL)
}
\arguments{
  \item{x}{a vector of numeric or ordered values}

  \item{na.rm}{ignored}
}
\value{
  the minimum value of the vector
}
\description{
  Compute the Minumum Value of a Vector
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

