\name{param.default}
\alias{param.default}
\title{Default method for param}
\usage{param.default(obj, num, ...)}
\description{If no \code{param} function is set for a Zelig model, then
this function will return NULL.}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{obj}{ignored parameter}
\item{num}{ignored parameter}
\item{...}{ignored parameters}}
