\name{param.logit}
\alias{param.logit}
\title{Param Method for the 'logit' Zelig Model...}
\usage{param.logit(obj, num, ...)}
\description{Param Method for the 'logit' Zelig Model}
\note{This method is used by the 'logit' Zelid model}
\value{a list to be cast as a 'parameters' object}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{obj}{a 'zelig' object}
\item{num}{an integer specifying the number of simulations to sample}
\item{...}{ignored parameters}}
