\name{parseFormula.formula}
\alias{parseFormula.formula}
\title{Parse Standard Formulae...}
\usage{\method{parseFormula}{formula}(obj, data=NULL)}
\description{Parse Standard Formulae}
\details{This method parses a formula-style Zelig formula}
\value{an object of type "parseFormula"}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{obj}{a formula}
\item{data}{a data frame}}
