\name{parseFormula.formula}
\alias{parseFormula.formula}
\title{Parse Standard Formulae}
\usage{
  \method{parseFormula}{formula}(obj, data=NULL)
}
\arguments{
  \item{obj}{a formula}

  \item{data}{a data frame}
}
\value{
  an object of type "parseFormula"
}
\description{
  This method parses a formula-style Zelig formula
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

