\name{parseFormula.list}
\alias{parseFormula.list}
\title{Parse List-Style Zelig Formulae}
\usage{
  \method{parseFormula}{list}(obj, data=NULL)
}
\arguments{
  \item{obj}{a list of formulae}

  \item{data}{a data frame}
}
\value{
  an object of type "parseFormula"
}
\description{
  This method parses a list-style Zelig formula.
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

