\name{plot.MI.sim}
\alias{plot.MI.sim}
\title{Plot graphs of simulated multiply-imputed data}
\usage{
  \method{plot}{MI.sim}(...)
}
\arguments{
  \item{...}{ignored parameters}
}
\value{
  NULL (invisibly)
}
\description{
  This function is currently unimplemented, and reserved
  for future use.
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

