\name{plot.MI.sim}
\alias{plot.MI.sim}
\title{Plot graphs of simulated multiply-imputed data...}
\usage{plot.MI.sim(...)}
\description{Plot graphs of simulated multiply-imputed data}
\details{This function is currently unimplemented, and reserved for future use.}
\value{NULL (invisibly)}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{...}{ignored parameters}}
