\name{plot.sim}
\alias{plot.sim}
\title{Method for plotting simulations...}
\usage{plot.sim(x, xlab="", ...)}
\description{Method for plotting simulations}
\value{the current graphical parameters. This is subject to change in
future implementations of Zelig}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{x}{a `sim' object}
\item{xlab}{labels for the x-axis}
\item{...}{parameters to be passed to the `truehist' function
which is implicitly called for numeric simulations}}
