\name{plot.sim}
\alias{plot.sim}
\title{Method for plotting simulations}
\usage{
  \method{plot}{sim}(x, ...)

  \method{plot}{sim}(x, ...)
}
\arguments{
  \item{x}{a `sim' object}

  \item{...}{parameters to be passed to the `truehist'
  function which is implicitly called for numeric
  simulations}

  \item{x}{a `sim' object}

  \item{...}{parameters to be passed to the `truehist'
  function which is implicitly called for numeric
  simulations}
}
\value{
  nothing

  nothing
}
\description{
  Plot simulated quantities of interest.

  Plot simulated quantities of interest.
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}

  Matt Owen \email{mowen@iq.harvard.edu}
}

