\name{print.qi}
\alias{print.qi}
\title{Print a Quantity of Interest in Human-Readable Form}
\usage{
  \method{print}{qi}(x, ...)
}
\arguments{
  \item{x}{a qi object}

  \item{...}{ignored parameters}
}
\value{
  the object that was printed (invisibly)
}
\description{
  Print simulated quantities of interest in a
  human-readable form
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

