\name{print.qi}
\alias{print.qi}
\title{Print a Quantity of Interest in Human-Readable Form...}
\usage{print.qi(x, ...)}
\description{Print a Quantity of Interest in Human-Readable Form}
\value{the object that was printed (invisibly)}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{x}{a qi object}
\item{...}{ignored parameters}}
