\name{print.qi.summarized}
\alias{print.qi.summarized}
\title{Print Method for Summarized Quantities of Interest...}
\usage{\method{print}{qi.summarized}(x, \dots)}
\description{Print Method for Summarized Quantities of Interest}
\value{x (invisibly)}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\seealso{\link{special_prints_MATRIX} and
\link{special_prints_LIST}}
\arguments{\item{x}{a 'summarized.qi' object}
\item{...}{parameters to be passed to the specific print functions}}
