\name{print.setx.mi}
\alias{print.setx.mi}
\title{Print a Bundle of Data-sets}
\usage{
  \method{print}{setx.mi}(x, ...)
}
\arguments{
  \item{x}{a \code{setx} object to print}

  \item{...}{ignored parameters}
}
\value{
  the \code{setx} object (invisibly)
}
\description{
  Print a Bundle of Data-sets
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

