\name{print.setx.mi}
\alias{print.setx.mi}
\title{Print a Bundle of Data-sets...}
\usage{\method{print}{setx.mi}(x, ...)}
\description{Print a Bundle of Data-sets}
\value{the \code{setx} object (invisibly)}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{x}{a \code{setx} object to print}
\item{...}{ignored parameters}}
