\name{print.setx}
\alias{print.setx}
\title{Print values of `setx' objects...}
\usage{print.setx(x, ...)}
\description{Print values of `setx' objects}
\value{the value of x (invisibly)}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{x}{a `setx' object}
\item{...}{ignored parameters}}
