\name{print.sim}
\alias{print.sim}
\title{Print values of `sim' objects...}
\usage{print.sim(x, ...)}
\description{Print values of `sim' objects
This function is currently unimplemented, and
included for future development}
\value{NULL (invisibly)}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{x}{a `sim' object (ignored)}
\item{...}{ignored parameters}}
