\name{print.summary.relogit}
\alias{print.summary.relogit}
\title{Print Summary of a Rare-event Logistic Model}
\usage{
  print.summary.relogit(x,
    digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
  \item{x}{...}

  \item{digits}{...}
}
\value{
  x (invisibly)
}
\description{
  ...
}

