\name{print.summary.relogit}
\alias{print.summary.relogit}
\title{Print Summary of a Rare-event Logistic Model}
\usage{
  \method{print}{summary.relogit}(x, digits = max(3,
    getOption("digits") - 3), ...)
}
\arguments{
  \item{x}{an ``relogit.summary'' object produced by the
  ``summary'' method.}

  \item{digits}{an integer specifying the number of digits
  of precision to specify}

  \item{...}{}
}
\value{
  x (invisibly)
}
\description{
  Prints the
}

