\name{print.summary.relogit2}
\alias{print.summary.relogit2}
\title{Print Summary of a Rare-event Logistic Model}
\usage{
  \method{print}{summary.relogit}(x, digits = max(3,
    getOption("digits") - 3), ...)
}
\arguments{
  \item{x}{the object to print}

  \item{digits}{an integer specifying the number of digits
  of precision}

  \item{...}{ignored parameters}
}
\value{
  x (invisibly)
}
\description{
  ...
}

