\name{print.summary.sim}
\alias{print.summary.sim}
\title{Print Values of a Summarized 'sim' Object}
\usage{
  print.summary.sim(x, ...)
}
\arguments{
  \item{x}{a 'summary.sim' object}

  \item{...}{ignored parameters}
}
\value{
  the value of the `summary.sim' object (invisibly)
}
\description{
  Print Values of a Summarized 'sim' Object
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

