\name{print.summary.sim}
\alias{print.summary.sim}
\title{Print Values of a Summarized 'sim' Object...}
\usage{print.summary.sim(x, ...)}
\description{Print Values of a Summarized 'sim' Object}
\value{the value of the `summary.sim' object (invisibly)}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{x}{a 'summary.sim' object}
\item{...}{ignored parameters}}
