\name{print.zelig}
\alias{print.zelig}
\title{Print values of ``zelig'' objects}
\usage{
  \method{print}{zelig}(x, ...)
}
\arguments{
  \item{x}{a `zelig' object}

  \item{...}{ignored parameters}
}
\value{
  the `zelig' object (invisibly)
}
\description{
  Print the zelig object as a list
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

