\name{print.zelig}
\alias{print.zelig}
\title{Print values of `zelig' objects}
\usage{
  print.zelig(x, ...)
}
\arguments{
  \item{x}{a `zelig' object}

  \item{...}{ignored parameters}
}
\value{
  the `zelig' object (invisibly)
}
\description{
  Print values of `zelig' objects
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

