\name{register.default}
\alias{register.default}
\title{Construct a list of available generic methods available for the
object used to fit the Zelig model}
\usage{
  register.default(obj)
}
\arguments{
  \item{obj}{a `zelig' object}
}
\value{
  a character vector containing names of S3 methods for the
  statistical model
}
\description{
  Construct a list of available generic methods available
  for the object used to fit the Zelig model
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

