\name{repl.default}
\alias{repl.default}
\title{Default Method for Replicating Statistics...}
\usage{repl.default(object, data, ...)}
\description{Default Method for Replicating Statistics}
\value{a replicated object}
\author{Kosuke Imai and Olivia Lau \email{mowen@iq.harvard.edu}}
\arguments{\item{object}{an object to replicate}
\item{data}{a data.frame}
\item{...}{ignored parameters}}
