\name{repl.default}
\alias{repl.default}
\title{Default Method for Replicating Statistics}
\usage{
  repl.default(object, data = NULL, ...)
}
\arguments{
  \item{object}{an object to replicate}

  \item{data}{a data.frame}

  \item{...}{ignored parameters}
}
\value{
  a replicated object
}
\description{
  Default Method for Replicating Statistics
}
\author{
  Kosuke Imai and Olivia Lau \email{mowen@iq.harvard.edu}
}

