\name{repl}
\alias{repl}
\title{Generic Method for Replicating Data}
\usage{
  repl(object, ...)
}
\arguments{
  \item{object}{a 'zelig' object}

  \item{...}{parameters}
}
\value{
  a replicated object
}
\description{
  Generic Method for Replicating Data
}
\author{
  Kosuke Imai and Olivia Lau \email{mowen@iq.harvard.edu}
}

