\name{simulations.parameters}
\alias{simulations.parameters}
\title{Extract Simulations from 'parmaeters' Objects...}
\usage{simulations.parameters(object, ...)}
\description{Extract Simulations from 'parmaeters' Objects}
\note{This function may not differ at all from coef.default}
\value{simulations, specified by the Zelig model, of
the ancillary parameters}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{object}{a 'parameters' object}
\item{\dots}{ignored}}
