\name{special_print_LIST}
\title{Method for Printing Summarized QI's in a List Form...}
\usage{.print.qi.summarized.LIST(x, ...)}
\description{Method for Printing Summarized QI's in a List Form}
\alias{special_print_LIST}
\alias{.print.qi.summarized.LIST}
\note{This function is used internall by Zelig}
\value{x (invisibly)}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{x}{a 'summarized.qi' object}
\item{...}{additional parameters to be used by the 'print.matrix' method}}
