\name{special_print_MATRIX}
\title{Method for Printing Summarized QI's in a Matrix Form...}
\usage{.print.qi.summarized.MATRIX(x, ...)}
\description{Method for Printing Summarized QI's in a Matrix Form}
\alias{special_print_MATRIX}
\alias{.print.qi.summarized.MATRIX}
\note{This function is used internall by Zelig}
\value{x (invisibly)}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{x}{a 'summarized.qi' object}
\item{...}{additional parameters}}
