\name{structuralToReduced}
\alias{structuralToReduced}
\title{Transform the Multilevel's Structural Formulas Into Reduced Form...}
\usage{structuralToReduced(f)}
\description{Transform the Multilevel's Structural Formulas Into Reduced Form}
\value{a formula in reduced form}
\author{Ferdinand Alimadhi, Kosuke Imai, and Olivia Lau}
\arguments{\item{f}{a list of formulas}}
