\name{summarize.default}
\alias{summarize.default}
\title{Summarize Simualted Quantities of Interest}
\usage{
  \method{summarize}{default}(obj)
}
\arguments{
  \item{obj}{a \code{qi} object, storing simulations of
  quantities of interest}
}
\value{
  a 'summarized.qi' object
}
\description{
  Summarize Simualted Quantities of Interest
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

