\name{summarize.default}
\alias{summarize.default}
\title{Summarize Simualted Quantities of Interest...}
\usage{\method{summarize}{default}(obj)}
\description{Summarize Simualted Quantities of Interest}
\value{a 'summarized.qi' object}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{obj}{a \code{qi} object, storing simulations of quantities of interest}}
