\name{summarize}
\alias{summarize}
\title{Generic methonf for summarizing simualted quantities of interest}
\usage{
  summarize(obj)
}
\arguments{
  \item{obj}{a \code{qi} object, storing simulations of
  quantities of interest}
}
\value{
  a \code{summarized.qi} object
}
\description{
  Generic methonf for summarizing simualted quantities of
  interest
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

