\name{summarize}
\alias{summarize}
\title{Generic methonf for summarizing simualted quantities of interest...}
\usage{summarize(qi)}
\description{Generic methonf for summarizing simualted quantities of interest}
\value{a `summarized.qi' object}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{qi}{a `qi' object, storing simulations of quantities of interest}}
