\name{summary.MI}
\alias{summary.MI}
\title{Summarry of Multiply Imputed Statistical Models...}
\usage{summary.MI(object, ...)}
\description{Summarry of Multiply Imputed Statistical Models}
\value{a list of summaries}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{object}{a set of fitted statistical models}
\item{...}{parameters to forward}}
