\name{summary.MI}
\alias{summary.MI}
\title{Summarry of Multiply Imputed Statistical Models}
\usage{
  \method{summary}{MI}(object, subset = NULL, ...)
}
\arguments{
  \item{object}{a set of fitted statistical models}

  \item{subset}{an integer vector, specifying the indices of the data.frames to be used in the subset}

  \item{...}{parameters to forward}
}
\value{
  a list of summaries
}
\description{
  ...
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

