\name{summary.MI}
\alias{summary.MI}
\title{Summarry of Multiply Imputed Statistical Models}
\usage{
  \method{summary}{MI}(object, ...)
}
\arguments{
  \item{object}{a set of fitted statistical models}

  \item{...}{parameters to forward}
}
\value{
  a list of summaries
}
\description{
  ...
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

