\name{summary.MI.sim}
\alias{summary.MI.sim}
\title{Method for summarizing simulations of multiply imputed quantities of interest}
\usage{
  \method{summary}{MI.sim}(object, ...)
}
\arguments{
  \item{object}{a `MI.sim' object}

  \item{...}{ignored parameters}
}
\value{
  a `summarized.MI.sim' object
}
\description{
  Method for summarizing simulations of multiply imputed
  quantities of interest
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

