\name{summary.sim}
\alias{summary.sim}
\title{Method for summarizing simulations of quantities of interest...}
\usage{summary.sim(object, ...)}
\description{Method for summarizing simulations of quantities of interest}
\value{a 'summarized.MI.sim' object}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{object}{a 'MI.sim' object}
\item{...}{ignored parameters}}
