\name{summary.zelig}
\alias{summary.zelig}
\title{Zelig Object Summaries}
\usage{
  \method{summary}{zelig}(object, ...)
}
\arguments{
  \item{object}{a zelig object}

  \item{...}{parameters forwarded to the generic summary
  object}
}
\value{
  the summary of the fitted model
}
\description{
  Compute summary data for zelig objects
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}
}

