\name{summary.zelig}
\alias{summary.zelig}
\title{Zelig Object Summaries...}
\usage{summary.zelig(object, ...)}
\description{Zelig Object Summaries
Compute summary data for zelig objects}
\value{the summary of the fitted model}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{object}{a zelig object}
\item{...}{parameters forwarded to the generic summary object}}
