\name{terms.multiple}
\alias{terms.multiple}
\title{Extract Terms from Multiple-component Formulas...}
\usage{\method{terms}{multiple}(x, data=NULL, ...)}
\description{Extract Terms from Multiple-component Formulas}
\value{the terms attribute}
\author{Kosuke Imai and Olivia Lau}
\arguments{\item{x}{a formula or list of formulas object}
\item{data}{a data.frame or NULL}
\item{...}{ignored parameters}}
