\name{terms.vglm}
\alias{terms.vglm}
\title{Model Terms for 'vglm' Models}
\usage{
  \method{terms}{vglm}(x, ...)
}
\arguments{
  \item{x}{a fitted model object from the VGAM library}

  \item{...}{ignored parameters}
}
\value{
  the models terms of this fitted model object
}
\description{
  Model Terms for 'vglm' Models
}
\author{
  Ferdinand Alimadhi, Kosuke Imai and Olivia Lau
}

