\name{terms.zelig}
\alias{terms.zelig}
\title{Model Terms for a Zelig Object}
\usage{
  \method{terms}{zelig}(x, ...)
}
\arguments{
  \item{x}{a \code{zelig} object}

  \item{...}{forwarded parameters}
}
\value{
  terms of the original fitted model
}
\description{
  This method simply extracts the model terms for the
  fitted model passed to the \code{zelig} function.
}

