\name{termsFromFormula}
\alias{termsFromFormula}
\title{Extract Terms from Zelig-style Formulae...}
\usage{termsFromFormula(obj)}
\description{Extract Terms from Zelig-style Formulae}
\details{This method is a sugary function to extract terms from any type of 
Zelig-style formula.}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{obj}{}}
