\name{TexCite}
\alias{TexCite}
\title{Get a TeX-style Citation}
\usage{
  TexCite(model)
}
\arguments{
  \item{model}{a character-string specifying the name of
  the Zelig model of which to describe in TeX-style}
}
\value{
  a string to be rendered as part of a LaTeX-style document
}
\description{
  Get a TeX-style Citation
}

