\name{toBuildFormula}
\alias{toBuildFormula}
\title{Build Formula ?}
\usage{toBuildFormula(Xnames, sepp="+")}
\description{Build Formula ???}
\details{This function builds a formula}
\value{a character-string}
\author{???}
\arguments{\item{Xnames}{a character-vector}
\item{sepp}{a seperator (???)}}
