\name{toString.call}
\alias{toString.call}
\title{Convert call Object to a String...}
\usage{\method{toString}{call}(x, ...)}
\description{Convert \code{call} Object to a String}
\details{This method concerts \code{call} objects into a simple, intuitive 
human-readable form.}
\value{a character-string representing the \code{call} object}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{x}{a \code{call} object}
\item{...}{ignored parameters}}
