\name{ucfirst}
\alias{ucfirst}
\title{Uppercase First Letter of a String...}
\usage{ucfirst(str)}
\description{Uppercase First Letter of a String}
\details{This method sets the first character of a string to its uppercase,
sets all other characters to lowercase.}
\value{a vector of character strings}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{str}{a vector of charaqcter-strings}}
