\name{user.prompt}
\alias{user.prompt}
\title{Prompt User}
\usage{
  user.prompt(msg = NULL)
}
\arguments{
  \item{msg}{a character-string, specifying a message to be
  displayed}
}
\value{
  This function is used for its side effects
}
\description{
  Prompts user to hit enter
}
\note{
  This function is primarily used by Zelig demo scripts
}

