\name{zelig2gamma}
\alias{zelig2gamma}
\title{Interface between gamma model and Zelig...}
\usage{zelig2gamma(formula, ..., data)}
\description{Interface between gamma model and Zelig
This function is exclusively for use by the `zelig' function}
\value{a list to be coerced into a zelig.call object}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{formula}{a formula}
\item{...}{ignored parameters}
\item{data}{a data.frame}}
