\name{zelig2negbinom}
\alias{zelig2negbinom}
\title{Interface between negbinom model and Zelig...}
\usage{zelig2negbinom(formula, weights, ..., data)}
\description{Interface between negbinom model and Zelig
This function is exclusively for use by the `zelig' function}
\value{a list to be coerced into a zelig.call object}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{formula}{a formula}
\item{weights}{a numeric vector}
\item{...}{ignored parameters}
\item{data}{a data.frame}}
