\name{zelig2relogit}
\alias{zelig2relogit}
\title{Zelig2 bridge function}
\usage{
  zelig2relogit(formula, ..., tau = NULL,
    bias.correct = NULL, case.control = NULL, data)
}
\arguments{
  \item{formula}{a formula object}

  \item{...}{ignored parameters}

  \item{tau}{...}

  \item{bias.correct}{...}

  \item{case.control}{...}

  \item{data}{a data.frame that will be used to fit the
  model}
}
\value{
  a list used internally by zelig
}
\description{
  ...
}
\note{
  T
}

