\name{ZeligDescribeModel}
\alias{ZeligDescribeModel}
\title{Produce a 'description' Object from the Name of a Model...}
\usage{ZeligDescribeModel(model.name)}
\description{Produce a 'description' Object from the Name of a Model}
\note{The 'description' object is a list-style object containing citation
information}
\value{a 'description' object specified by the 'model.name' parameter. This
object is created by executing the specified Zelig models' 'describe'
function}
\arguments{\item{model.name}{a character-string specifying a Zelig model}}
